refactor: post.controller
