+++ src/controllers/post/index.js
@@ -1,0 +1,4 @@
+export * from './create.controllers.js';
+export * from './read.controllers.js';
+export * from './update.controllers.js';
+export * from './delete.controllers.js';
